+++ vectorizers/__init__.py
@@ -1,13 +1,13 @@
-from ._vectorizers import TokenCooccurrenceVectorizer
+from .token_cooccurrence_vectorizer import TokenCooccurrenceVectorizer
 from ._vectorizers import DistributionVectorizer
 from ._vectorizers import HistogramVectorizer
-from ._vectorizers import SkipgramVectorizer
-from ._vectorizers import NgramVectorizer
+from .skip_gram_vectorizer import SkipgramVectorizer
+from .ngram_vectorizer import NgramVectorizer
 from .kde_vectorizer import KDEVectorizer
 from ._vectorizers import ProductDistributionVectorizer
 from ._vectorizers import Wasserstein1DHistogramTransformer
 from ._vectorizers import SequentialDifferenceTransformer
-from ._vectorizers import LabelledTreeCooccurrenceVectorizer
+from .tree_token_cooccurrence import LabelledTreeCooccurrenceVectorizer
 
 from .utils import cast_tokens_to_strings
